+++ src/components/manageProfile/profileAndVisibility/AboutYouComponent.js
@@ -1,5 +1,4 @@
-class ManageProfileComponent {
-    
+class AboutYouComponent {
     get permissionsOptions() { return $('div[data-test-selector=visibility-wrapper-localtime]'); }
     get permissionsAnyone() { return $('div=Anyone'); }
 
@@ -9,14 +8,8 @@ class ManageProfileComponent {
         await this.permissionsOptions.click();
         await this.permissionsAnyone.waitForDisplayed();
         await this.permissionsAnyone.click();
-    }
-
-    async switchToWindow() {
-        await browser.pause(2000);
-        const handles = await browser.getWindowHandles();
-        await browser.switchToWindow(handles[1]);
-    }
-
+    }    
+    
 }
 
-module.exports = new ManageProfileComponent();
+module.exports = new AboutYouComponent();
